Prepare ApplicationResult view
